3.108.1085 * Add Server status * Shorten  RTCM1005 and 1033 * Add RTK message count * Add temp to power managment page * Add GPS to GPS page

+++ WSRtkServer/include/Web/WebPageWrapper.h
@@ -172,7 +172,7 @@ public:
 	{
 		TableRow(indent, name, value.c_str());
 	}
-	void TableRow(int indent, const std::string &name, int32_t value)
+	void TableRow(int indent, const std::string &name, int64_t value)
 	{
 		TableRow(indent, name, ToThousands(value).c_str(), true);
 	}
